Standardized test
- Tests are now all unit, because interface; duh.
- FQNs fixed and standardized.
- Removed functionality test.
- Corrected error wording.
